+++ onegov/town/tests/test_forms.py
@@ -1,12 +1,14 @@
 import pytest
 
-from datetime import date, datetime, time
+from datetime import date, datetime, time, timedelta
 from dateutil.rrule import MO, WE
 from libres.db.models import Allocation
+from onegov.core.utils import Bunch
 from onegov.town.forms import (
     DaypassAllocationForm,
     ReservationForm,
-    RoomAllocationForm
+    RoomAllocationForm,
+    RoomAllocationEditForm
 )
 from onegov.town.forms.allocation import AllocationFormHelpers
 
@@ -187,6 +189,21 @@ def test_generate_datetimes():
     ]
 
 
+def test_combine_datetime():
+    helper = AllocationFormHelpers()
+
+    helper.date = Bunch(data=date(2015, 1, 1))
+    helper.time = Bunch(data=None)
+    assert helper.combine_datetime('date', 'time') == datetime(2015, 1, 1)
+
+    helper.time = Bunch(data='12:00')
+    assert helper.combine_datetime('date', 'time') == datetime(2015, 1, 1, 12)
+
+    helper.date = Bunch(data=None)
+    helper.time = Bunch(data=None)
+    assert helper.combine_datetime('date', 'time') is None
+
+
 def test_reservation_form_partly_available():
     allocation = Allocation()
 
@@ -234,3 +251,36 @@ def test_reservation_form_quota():
     allocation.quota_limit = 0
     form = ReservationForm.for_allocation(allocation)()
     assert hasattr(form, 'quota')
+
+
+def test_edit_room_alllocation_form():
+
+    form = RoomAllocationEditForm()
+    form.apply_dates(datetime(2015, 1, 1, 12, 0), datetime(2015, 1, 1, 18, 0))
+
+    assert form.date.data == date(2015, 1, 1)
+    assert form.start_time.data == '12:00'
+    assert form.end_time.data == '18:00'
+
+    form.as_whole_day.data = 'no'
+    assert form.dates == (
+        datetime(2015, 1, 1, 12, 0),
+        datetime(2015, 1, 1, 18, 0)
+    )
+
+    form.as_whole_day.data = 'yes'
+    assert form.dates == (
+        datetime(2015, 1, 1),
+        datetime(2015, 1, 2) - timedelta(microseconds=1)
+    )
+
+    form.apply_model(Bunch(
+        display_start=lambda: datetime(2015, 1, 1, 12),
+        display_end=lambda: datetime(2015, 1, 1, 18),
+        whole_day=False
+    ))
+
+    assert form.dates == (
+        datetime(2015, 1, 1, 12, 0),
+        datetime(2015, 1, 1, 18, 0)
+    )